test(chat): isolate real network smokes
