+++ package.json
@@ -38,6 +38,7 @@
 		"typecheck": "node ./scripts/run-typechecks.mjs",
 		"build": "npm --workspaces --if-present run build",
 		"test": "node ./scripts/run-tests.mjs",
+		"smoke:chat-network": "node ./scripts/run-pi-chat-network-smoke.mjs",
 		"check": "npm run build && node ./scripts/run-checks.mjs",
 		"precommit": "npm run biome:check && npm run typecheck",
 		"prepare": "husky",
