+++ scripts/run-pi-chat-network-smoke.mjs
@@ -1,0 +1,28 @@
+#!/usr/bin/env node
+
+import { spawnSync } from "node:child_process";
+import fs from "node:fs";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+
+const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
+const outDir = path.join(root, "node_modules", ".cache", "pi-chat-network-smoke");
+const tsc = path.join(
+	root,
+	"node_modules",
+	".bin",
+	process.platform === "win32" ? "tsc.cmd" : "tsc",
+);
+
+fs.rmSync(outDir, { recursive: true, force: true });
+run(tsc, ["-p", "packages/pi-chat/tsconfig.network-smoke.json"]);
+run(process.execPath, [path.join(outDir, "packages", "pi-chat", "scripts", "network-smoke.js")]);
+
+function run(command, args) {
+	const result = spawnSync(command, args, { cwd: root, stdio: "inherit", env: process.env });
+	if (result.error) {
+		console.error(result.error.message);
+		process.exit(1);
+	}
+	if (result.status !== 0) process.exit(result.status ?? 1);
+}
